feat : IceWizard Attack logic, fix : Enemy HP Weight float notation

+++ DefenceGame/DefenceGame/GoldGoblin.cpp
@@ -8,7 +8,7 @@ GoldGoblin::GoldGoblin(ENTITY_TYPE type, std::string renderString, COLOR color, 
 	_type = type;
 	_renderString = renderString;
 	_color = color;
-	_hp = std::round((float)hp * (float)(1 + GET_SINGLETON(WaveManager)->getCurrentWave() / (float)40));
+	_hp = std::round((float)hp * (float)(1 + GET_SINGLETON(WaveManager)->getCurrentWave() / 40.0f));
 	_moveTime = moveTime;
 	_rewardGold = rewardGold;
 	_originColor = _color;
